fix(Pages.styles): move out of /pages, causing build failure
